commit add test project service

+++ task_poject/src/main/java/by/milavitsky/task_poject/dto/ProjectDTO.java
@@ -11,6 +11,7 @@ import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.Size;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 @NoArgsConstructor
@@ -30,7 +31,7 @@ public class ProjectDTO implements Serializable {
     private String projectDescription;
 
     @Positive(message = "Budget should be positive")
-    private Long budget;
+    private BigDecimal budget;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.s")
     private String dateOfStart;
